rag-chatbot/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
chromadb
langchain
langchain-community
langchain-core
langchain-text-splitters
sentence-transformers
torch
pypdf
python-docx
unstructured

requests
PyPDF2
docx2txt
pillow
transformers
llama-cpp-python